Update ACMain.c
Added a couple lines so that if the mechanical revolutions per second is less than 0.5, then there would be no regen allowed.  When driving down the street, you don't want to come to a stop, and then suddenly have the motor reverse direction, making you fly into reverse, because of the "regen".  LOL.